requirements.txt
beautifulsoup4
ergast-py
lxml
numpy
pandas
requests
streamlit
tqdm
PyYAML